+++ tests/test_tokenizer.py
@@ -1,5 +1,5 @@
 import io
-from tokenize import NEWLINE, NUMBER, TokenInfo, generate_tokens
+from tokenize import NEWLINE, NUMBER, ENDMARKER, TokenInfo, generate_tokens
 
 from pegen.tokenizer import Tokenizer
 
@@ -31,3 +31,12 @@ def test_last_non_whitespace():
     assert t.getnext() == TokenInfo(NUMBER, "1", (2, 0), (2, 1), "1\n")
     assert t.getnext() == TokenInfo(NEWLINE, "\n", (2, 1), (2, 2), "1\n")
     assert t.get_last_non_whitespace_token() == TokenInfo(NUMBER, "1", (2, 0), (2, 1), "1\n")
+
+
+def test_get_lines():
+    source = io.StringIO("1\n2\n3")
+    t = Tokenizer(generate_tokens(source.readline))
+    while True:
+        if t.getnext().type == ENDMARKER:
+            break
+    assert t.get_lines([1, 2, 3]) == ["1\n", "2\n", "3"]